+++ src/Symfony/Component/Security/Core/Tests/Authorization/Strategy/ConsensusStrategyTest.php
@@ -9,9 +9,11 @@
  * file that was distributed with this source code.
  */
 
-namespace Authorization\Strategy;
+namespace Symfony\Component\Security\Core\Tests\Authorization\Strategy;
 
+use Symfony\Component\Security\Core\Authorization\AccessDecision;
 use Symfony\Component\Security\Core\Authorization\Strategy\ConsensusStrategy;
+use Symfony\Component\Security\Core\Authorization\Voter\Vote;
 use Symfony\Component\Security\Core\Test\AccessDecisionStrategyTestCase;
 
 class ConsensusStrategyTest extends AccessDecisionStrategyTestCase
@@ -20,21 +22,57 @@ class ConsensusStrategyTest extends AccessDecisionStrategyTestCase
     {
         $strategy = new ConsensusStrategy();
 
-        yield [$strategy, self::getVoters(1, 0, 0), true];
-        yield [$strategy, self::getVoters(1, 2, 0), false];
-        yield [$strategy, self::getVoters(2, 1, 0), true];
-        yield [$strategy, self::getVoters(0, 0, 1), false];
+        yield [$strategy, self::getVoters(1, 0, 0), AccessDecision::createGranted([
+            Vote::createGranted(),
+        ])];
+        yield [$strategy, self::getVoters(1, 2, 0), AccessDecision::createDenied([
+            Vote::createGranted(),
+            Vote::createDenied(),
+            Vote::createDenied(),
+        ])];
+        yield [$strategy, self::getVoters(2, 1, 0), AccessDecision::createGranted([
+            Vote::createGranted(),
+            Vote::createGranted(),
+            Vote::createDenied(),
+        ])];
+        yield [$strategy, self::getVoters(0, 0, 1), AccessDecision::createDenied([
+            Vote::createAbstain(),
+        ])];
 
-        yield [$strategy, self::getVoters(2, 2, 0), true];
-        yield [$strategy, self::getVoters(2, 2, 1), true];
+        yield [$strategy, self::getVoters(2, 2, 0), AccessDecision::createGranted([
+            Vote::createGranted(),
+            Vote::createGranted(),
+            Vote::createDenied(),
+            Vote::createDenied(),
+        ])];
+        yield [$strategy, self::getVoters(2, 2, 1), AccessDecision::createGranted([
+            Vote::createGranted(),
+            Vote::createGranted(),
+            Vote::createDenied(),
+            Vote::createDenied(),
+            Vote::createAbstain(),
+        ])];
 
         $strategy = new ConsensusStrategy(true);
 
-        yield [$strategy, self::getVoters(0, 0, 1), true];
+        yield [$strategy, self::getVoters(0, 0, 1), AccessDecision::createGranted([
+            Vote::createAbstain(),
+        ])];
 
         $strategy = new ConsensusStrategy(false, false);
 
-        yield [$strategy, self::getVoters(2, 2, 0), false];
-        yield [$strategy, self::getVoters(2, 2, 1), false];
+        yield [$strategy, self::getVoters(2, 2, 0), AccessDecision::createDenied([
+            Vote::createGranted(),
+            Vote::createGranted(),
+            Vote::createDenied(),
+            Vote::createDenied(),
+        ])];
+        yield [$strategy, self::getVoters(2, 2, 1), AccessDecision::createDenied([
+            Vote::createGranted(),
+            Vote::createGranted(),
+            Vote::createDenied(),
+            Vote::createDenied(),
+            Vote::createAbstain(),
+        ])];
     }
 }