+++ src/Symfony/Component/Security/Http/Tests/Firewall/AccessListenerTest.php
@@ -20,6 +20,7 @@ use Symfony\Component\Security\Core\Authentication\Token\NullToken;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorage;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
+use Symfony\Component\Security\Core\Authorization\AccessDecision;
 use Symfony\Component\Security\Core\Authorization\AccessDecisionManagerInterface;
 use Symfony\Component\Security\Core\Authorization\Voter\AuthenticatedVoter;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
@@ -32,6 +33,58 @@ class AccessListenerTest extends TestCase
 {
     public function testHandleWhenTheAccessDecisionManagerDecidesToRefuseAccess()
     {
+        $request = new Request();
+
+        $accessMap = $this->createMock(AccessMapInterface::class);
+        $accessMap
+            ->expects($this->any())
+            ->method('getPatterns')
+            ->with($this->equalTo($request))
+            ->willReturn([['foo' => 'bar'], null])
+        ;
+
+        $token = new class() extends AbstractToken {
+            public function getCredentials(): mixed
+            {
+            }
+        };
+
+        $tokenStorage = $this->createMock(TokenStorageInterface::class);
+        $tokenStorage
+            ->expects($this->any())
+            ->method('getToken')
+            ->willReturn($token)
+        ;
+
+        $accessDecisionManager = $this
+            ->getMockBuilder(AccessDecisionManagerInterface::class)
+            ->onlyMethods(['decide'])
+            ->addMethods(['getDecision'])
+            ->getMock();
+        $accessDecisionManager
+            ->expects($this->once())
+            ->method('getDecision')
+            ->with($this->equalTo($token), $this->equalTo(['foo' => 'bar']), $this->equalTo($request))
+            ->willReturn(AccessDecision::createDenied())
+        ;
+
+        $listener = new AccessListener(
+            $tokenStorage,
+            $accessDecisionManager,
+            $accessMap
+        );
+
+        $this->expectException(AccessDeniedException::class);
+
+        $listener(new RequestEvent($this->createMock(HttpKernelInterface::class), $request, HttpKernelInterface::MAIN_REQUEST));
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testHandleWhenTheAccessDecisionManagerDecidesToRefuseAccessLegacy()
+    {
+        $this->expectException(AccessDeniedException::class);
         $request = new Request();
 
         $accessMap = $this->createMock(AccessMapInterface::class);
@@ -142,11 +195,15 @@ class AccessListenerTest extends TestCase
             ->willReturn([['foo' => 'bar'], null])
         ;
 
-        $accessDecisionManager = $this->createMock(AccessDecisionManagerInterface::class);
+        $accessDecisionManager = $this
+            ->getMockBuilder(AccessDecisionManagerInterface::class)
+            ->onlyMethods(['decide'])
+            ->addMethods(['getDecision'])
+            ->getMock();
         $accessDecisionManager->expects($this->once())
-            ->method('decide')
+            ->method('getDecision')
             ->with($this->isInstanceOf(NullToken::class))
-            ->willReturn(false);
+            ->willReturn(AccessDecision::createDenied());
 
         $listener = new AccessListener(
             $tokenStorage,
@@ -172,11 +229,15 @@ class AccessListenerTest extends TestCase
             ->willReturn([[AuthenticatedVoter::PUBLIC_ACCESS], null])
         ;
 
-        $accessDecisionManager = $this->createMock(AccessDecisionManagerInterface::class);
+        $accessDecisionManager = $this
+            ->getMockBuilder(AccessDecisionManagerInterface::class)
+            ->onlyMethods(['decide'])
+            ->addMethods(['getDecision'])
+            ->getMock();
         $accessDecisionManager->expects($this->once())
-            ->method('decide')
+            ->method('getDecision')
             ->with($this->isInstanceOf(NullToken::class), [AuthenticatedVoter::PUBLIC_ACCESS])
-            ->willReturn(true);
+            ->willReturn(AccessDecision::createGranted());
 
         $listener = new AccessListener(
             $tokenStorage,
@@ -202,11 +263,15 @@ class AccessListenerTest extends TestCase
             ->willReturn([[AuthenticatedVoter::PUBLIC_ACCESS], null])
         ;
 
-        $accessDecisionManager = $this->createMock(AccessDecisionManagerInterface::class);
+        $accessDecisionManager = $this
+            ->getMockBuilder(AccessDecisionManagerInterface::class)
+            ->onlyMethods(['decide'])
+            ->addMethods(['getDecision'])
+            ->getMock();
         $accessDecisionManager->expects($this->once())
-            ->method('decide')
+            ->method('getDecision')
             ->with($this->equalTo($token), [AuthenticatedVoter::PUBLIC_ACCESS])
-            ->willReturn(true);
+            ->willReturn(AccessDecision::createGranted());
 
         $listener = new AccessListener(
             $tokenStorage,
@@ -235,12 +300,16 @@ class AccessListenerTest extends TestCase
         $tokenStorage = new TokenStorage();
         $tokenStorage->setToken($authenticatedToken);
 
-        $accessDecisionManager = $this->createMock(AccessDecisionManagerInterface::class);
+        $accessDecisionManager = $this
+            ->getMockBuilder(AccessDecisionManagerInterface::class)
+            ->onlyMethods(['decide'])
+            ->addMethods(['getDecision'])
+            ->getMock();
         $accessDecisionManager
             ->expects($this->once())
-            ->method('decide')
+            ->method('getDecision')
             ->with($this->equalTo($authenticatedToken), $this->equalTo(['foo' => 'bar', 'bar' => 'baz']), $this->equalTo($request), true)
-            ->willReturn(true)
+            ->willReturn(AccessDecision::createGranted())
         ;
 
         $listener = new AccessListener(