+++ src/Symfony/Component/Security/Http/Tests/Firewall/SwitchUserListenerTest.php
@@ -20,6 +20,7 @@ use Symfony\Component\HttpKernel\HttpKernelInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorage;
 use Symfony\Component\Security\Core\Authentication\Token\SwitchUserToken;
 use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
+use Symfony\Component\Security\Core\Authorization\AccessDecision;
 use Symfony\Component\Security\Core\Authorization\AccessDecisionManagerInterface;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Security\Core\Exception\AuthenticationCredentialsNotFoundException;
@@ -45,7 +46,11 @@ class SwitchUserListenerTest extends TestCase
         $this->tokenStorage = new TokenStorage();
         $this->userProvider = new InMemoryUserProvider(['kuba' => []]);
         $this->userChecker = $this->createMock(UserCheckerInterface::class);
-        $this->accessDecisionManager = $this->createMock(AccessDecisionManagerInterface::class);
+        $this->accessDecisionManager = $this
+            ->getMockBuilder(AccessDecisionManagerInterface::class)
+            ->onlyMethods(['decide'])
+            ->addMethods(['getDecision'])
+            ->getMock();
         $this->request = new Request();
         $this->event = new RequestEvent($this->createMock(HttpKernelInterface::class), $this->request, HttpKernelInterface::MAIN_REQUEST);
     }
@@ -145,8 +150,8 @@ class SwitchUserListenerTest extends TestCase
         $this->request->query->set('_switch_user', 'kuba');
 
         $this->accessDecisionManager->expects($this->once())
-            ->method('decide')->with($token, ['ROLE_ALLOWED_TO_SWITCH'])
-            ->willReturn(false);
+            ->method('getDecision')->with($token, ['ROLE_ALLOWED_TO_SWITCH'])
+            ->willReturn(AccessDecision::createDenied());
 
         $listener = new SwitchUserListener($this->tokenStorage, $this->userProvider, $this->userChecker, 'provider123', $this->accessDecisionManager);
 
@@ -163,7 +168,7 @@ class SwitchUserListenerTest extends TestCase
         $this->request->query->set('_switch_user', 'not-existing');
 
         $this->accessDecisionManager->expects($this->never())
-            ->method('decide');
+            ->method('getDecision');
 
         $listener = new SwitchUserListener($this->tokenStorage, $this->userProvider, $this->userChecker, 'provider123', $this->accessDecisionManager);
 
@@ -180,13 +185,39 @@ class SwitchUserListenerTest extends TestCase
         $this->request->query->set('_switch_user', 'kuba');
 
         $this->accessDecisionManager->expects($this->once())
-            ->method('decide')->with($token, ['ROLE_ALLOWED_TO_SWITCH'], $this->callback(fn ($user) => 'kuba' === $user->getUserIdentifier()))
-            ->willReturn(true);
+            ->method('getDecision')->with($token, ['ROLE_ALLOWED_TO_SWITCH'], $this->callback(fn ($user) => 'kuba' === $user->getUserIdentifier()))
+            ->willReturn(AccessDecision::createGranted());
 
         $this->userChecker->expects($this->once())
             ->method('checkPostAuth')->with($this->callback(fn ($user) => 'kuba' === $user->getUserIdentifier()));
 
         $listener = new SwitchUserListener($this->tokenStorage, $this->userProvider, $this->userChecker, 'provider123', $this->accessDecisionManager);
+        $listener($this->event);
+
+        $this->assertSame([], $this->request->query->all());
+        $this->assertSame('', $this->request->server->get('QUERY_STRING'));
+        $this->assertInstanceOf(UsernamePasswordToken::class, $this->tokenStorage->getToken());
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testSwitchUserLegacy()
+    {
+        $token = new UsernamePasswordToken(new InMemoryUser('username', '', ['ROLE_FOO']), 'key', ['ROLE_FOO']);
+
+        $this->tokenStorage->setToken($token);
+        $this->request->query->set('_switch_user', 'kuba');
+
+        $accessDecisionManager = $this->createMock(AccessDecisionManagerInterface::class);
+        $accessDecisionManager->expects($this->once())
+            ->method('decide')->with($token, ['ROLE_ALLOWED_TO_SWITCH'], $this->callback(function ($user) { return 'kuba' === $user->getUserIdentifier(); }))
+            ->willReturn(true);
+
+        $this->userChecker->expects($this->once())
+            ->method('checkPostAuth')->with($this->callback(fn ($user) => 'kuba' === $user->getUserIdentifier()));
+
+        $listener = new SwitchUserListener($this->tokenStorage, $this->userProvider, $this->userChecker, 'provider123', $accessDecisionManager);
         $listener($this->event);
 
         $this->assertSame([], $this->request->query->all());
@@ -206,8 +237,8 @@ class SwitchUserListenerTest extends TestCase
 
         $targetsUser = $this->callback(fn ($user) => 'kuba' === $user->getUserIdentifier());
         $this->accessDecisionManager->expects($this->once())
-            ->method('decide')->with($originalToken, ['ROLE_ALLOWED_TO_SWITCH'], $targetsUser)
-            ->willReturn(true);
+            ->method('getDecision')->with($originalToken, ['ROLE_ALLOWED_TO_SWITCH'], $targetsUser)
+            ->willReturn(AccessDecision::createGranted());
 
         $this->userChecker->expects($this->once())
             ->method('checkPostAuth')->with($targetsUser);
@@ -232,8 +263,8 @@ class SwitchUserListenerTest extends TestCase
         $this->userProvider->createUser($user = new InMemoryUser('0', null));
 
         $this->accessDecisionManager->expects($this->once())
-            ->method('decide')->with($token, ['ROLE_ALLOWED_TO_SWITCH'])
-            ->willReturn(true);
+            ->method('getDecision')->with($token, ['ROLE_ALLOWED_TO_SWITCH'])
+            ->willReturn(AccessDecision::createGranted());
 
         $this->userChecker->expects($this->once())
             ->method('checkPostAuth')->with($this->callback(fn ($argUser) => $user->isEqualTo($argUser)));
@@ -259,8 +290,8 @@ class SwitchUserListenerTest extends TestCase
 
         $targetsUser = $this->callback(fn ($user) => 'kuba' === $user->getUserIdentifier());
         $this->accessDecisionManager->expects($this->once())
-            ->method('decide')->with($token, ['ROLE_ALLOWED_TO_SWITCH'], $targetsUser)
-            ->willReturn(true);
+            ->method('getDecision')->with($token, ['ROLE_ALLOWED_TO_SWITCH'], $targetsUser)
+            ->willReturn(AccessDecision::createGranted());
 
         $this->userChecker->expects($this->once())
             ->method('checkPostAuth')->with($targetsUser);
@@ -284,8 +315,8 @@ class SwitchUserListenerTest extends TestCase
         $this->request->query->set('_switch_user', 'kuba');
 
         $this->accessDecisionManager->expects($this->any())
-            ->method('decide')->with($token, ['ROLE_ALLOWED_TO_SWITCH'], $this->callback(fn ($user) => 'kuba' === $user->getUserIdentifier()))
-            ->willReturn(true);
+            ->method('getDecision')->with($token, ['ROLE_ALLOWED_TO_SWITCH'], $this->callback(fn ($user) => 'kuba' === $user->getUserIdentifier()))
+            ->willReturn(AccessDecision::createGranted());
 
         $dispatcher = $this->createMock(EventDispatcherInterface::class);
         $dispatcher
@@ -329,8 +360,8 @@ class SwitchUserListenerTest extends TestCase
 
         $targetsUser = $this->callback(fn ($user) => 'kuba' === $user->getUserIdentifier());
         $this->accessDecisionManager->expects($this->once())
-            ->method('decide')->with($token, ['ROLE_ALLOWED_TO_SWITCH'], $targetsUser)
-            ->willReturn(true);
+            ->method('getDecision')->with($token, ['ROLE_ALLOWED_TO_SWITCH'], $targetsUser)
+            ->willReturn(AccessDecision::createGranted());
 
         $this->userChecker->expects($this->once())
             ->method('checkPostAuth')->with($targetsUser);